Added basic DI Refactored database block for di
