+++ app/build.gradle
@@ -2,13 +2,14 @@ plugins {
     id 'com.android.application'
     id 'kotlin-android'
     id 'kotlin-kapt'
+    id 'dagger.hilt.android.plugin'
 }
 
 android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.myapplication"
+        applicationId "com.brxq.gyminstructor"
         minSdk 24
         targetSdk 31
         versionCode 1
@@ -69,5 +70,11 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.3.72"
     api "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.5"
     api "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.3.5"
+
+    // Dagger hilt
+    implementation 'com.google.dagger:hilt-android:2.41'
+    kapt 'com.google.dagger:hilt-compiler:2.41'
+
+
 
 }
