+++ app/src/main/java/com/brxq/gyminstructor/GymInstructorApp.kt
@@ -1,0 +1,7 @@
+package com.brxq.gyminstructor
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class GymInstructorApp : Application()
