+++ app/src/main/java/com/brxq/gyminstructor/MainActivity.kt
@@ -1,0 +1,34 @@
+package com.brxq.gyminstructor
+
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.navigation.findNavController
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        actionBar?.hide()
+
+        val bottomNav = findViewById<BottomNavigationView>(R.id.bottomNavigationView)
+        val navController = findNavController(R.id.fragment)
+
+        bottomNav.setupWithNavController(navController)
+
+
+    }
+
+}
+
+
+
+
+
+
