+++ app/src/main/java/com/brxq/gyminstructor/model/ProgramHasExercise.kt
@@ -1,0 +1,14 @@
+package com.brxq.gyminstructor.model
+
+import androidx.room.Embedded
+import androidx.room.Relation
+
+data class ProgramHasExercise(
+    @Embedded
+    val program : Program,
+    @Relation(
+        parentColumn = "program_id",
+        entityColumn = "program_id"
+    )
+    val listOfExercises : List<Exercise>
+)
