+++ app/src/main/java/com/brxq/gyminstructor/ui/exercise/ExerciseViewModel.kt
@@ -1,0 +1,7 @@
+package com.brxq.gyminstructor.ui.exercise
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+
+class ExerciseViewModel(application: Application) : AndroidViewModel(application) {
+}
