+++ app/src/main/java/com/brxq/gyminstructor/util/Constants.kt
@@ -1,0 +1,7 @@
+package com.brxq.gyminstructor.util
+
+object Constants {
+
+    const val EXERCISE_DATABASE = "exercise_database"
+
+}
