--- app/src/main/java/com/example/myapplication/MainActivity.kt
@@ -1,32 +1,0 @@
-package com.example.myapplication
-
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.navigation.findNavController
-import androidx.navigation.ui.setupWithNavController
-import com.google.android.material.bottomnavigation.BottomNavigationView
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        actionBar?.hide()
-
-        val bottomNav = findViewById<BottomNavigationView>(R.id.bottomNavigationView)
-        val navController = findNavController(R.id.fragment)
-
-        bottomNav.setupWithNavController(navController)
-
-
-    }
-
-}
-
-
-
-
-
-
